requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
flask-openapi3==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
Whoosh==2.7.4
